Custom color and direction for ElectricFieldPanel.
Currently color and direction are deactivated for other panels.
